// frontend/src/types.ts
export enum SoundStemCategory {
    DRUMS = 'Drums',
    BASS = 'Bass',
    MELODY = 'Melody',
    PADS = 'Pads',
    FX = 'FX',
}
// ...
export enum AssetType {
    IMAGE = 'Image',
    AUDIO = 'Audio',
    VIDEO = 'Video',
}

export interface Asset {
    id: string;
    name: string;
    type: AssetType;
    url: string;
    category?: SoundStemCategory;
}